Support both major elvish dialects
"Elvish" is insufficiently precise; the major dialects in Tolkien's
legendarium are Quenya, the old speech of the High Elves, and Sindarin,
the common tongue at the time of the Fellowship. It is very important
that this distinction be correctly represented here.

Although the correct default should be Sindarin, since that is the
elvish tongue in common use during the best documented period of
Middle-earth's history, we here retain the Quenya default for the sake
of backwards compatibility.