Support both major elvish dialects
"Elvish" is insufficiently precise; the major dialects in Tolkien's
legendarium are Quenya, the old speech of the High Elves, and Sindarin,
the common tongue at the time of the Fellowship. It is very important
that this distinction be correctly represented here. A deprecation
warning has been added for "elvish"; however, in this version it
continues to return the Quenya form, itself corrected to include diacritics.